+++ src/ServiceProvider.php
@@ -4,7 +4,7 @@ namespace AlphaSnow\AliyunOss;
 
 use AlphaSnow\AliyunOss\Plugins\PutRemoteFile;
 use Illuminate\Support\ServiceProvider as BaseServiceProvider;
-use League\Flysystem\Config;
+use League\Flysystem\Config as FlysystemConfig;
 use League\Flysystem\Filesystem;
 use OSS\OssClient;
 
@@ -30,23 +30,26 @@ class ServiceProvider extends BaseServiceProvider
     public function register()
     {
         $this->app->bind('aliyun-oss.oss-client', function ($app, array $config) {
-            $ossConfig = new AliyunOssConfig($config);
+            $ossConfig = new Config($config);
+
             $client = $app->make(OssClient::class, $ossConfig->getOssClientParameters());
             $client->setUseSSL($ossConfig->get('use_ssl', false));
             return $client;
         });
 
-        $this->app->bind('aliyun-oss.oss-filesystem', function ($app, array $config) {
+
+        $this->app->bind('aliyun-oss.oss-adapter', function ($app, array $config) {
             $client = $app->make('aliyun-oss.oss-client', $config);
-            $adapter = new AliyunOssAdapter($client, new AliyunOssConfig($config));
-            $filesystem = new Filesystem($adapter, new Config(['disable_asserts' => true]));
-            $filesystem->addPlugin(new PutRemoteFile());
-            return $filesystem;
+
+            return new Adapter($client, new Config($config));
         });
 
-        $this->app->singleton('aliyun-oss.client', function ($app) {
-            $config = $app->get('config')->get('filesystems.disks.aliyun');
-            return $app->make('aliyun-oss.oss-client', $config);
+        $this->app->bind('aliyun-oss.oss-filesystem', function ($app, array $config) {
+            $adapter = $app->make('aliyun-oss.oss-adapter', $config);
+
+            $filesystem = new Filesystem($adapter, new FlysystemConfig(['disable_asserts' => true]));
+            $filesystem->addPlugin(new PutRemoteFile());
+            return $filesystem;
         });
     }
 }