+++ tests/ConfigTest.php
@@ -2,7 +2,7 @@
 
 namespace AlphaSnow\AliyunOss\Tests;
 
-use AlphaSnow\AliyunOss\AliyunOssConfig;
+use AlphaSnow\AliyunOss\Config;
 
 class ConfigTest extends TestCase
 {
@@ -29,17 +29,17 @@ class ConfigTest extends TestCase
             'domain' => 'oss.my-domain.com',
             'use_domain_endpoint' => false,
         ];
-        $ossConfig = new AliyunOssConfig($config);
+        $ossConfig = new Config($config);
         $endpoint = $ossConfig->getOssEndpoint();
         $this->assertSame($config['endpoint'], $endpoint);
 
         $config['use_domain_endpoint'] = true;
-        $ossConfig = new AliyunOssConfig($config);
+        $ossConfig = new Config($config);
         $endpoint = $ossConfig->getOssEndpoint();
         $this->assertSame($config['domain'], $endpoint);
 
         $config['internal'] = 'oss-cn-shanghai-internal.aliyuncs.com';
-        $ossConfig = new AliyunOssConfig($config);
+        $ossConfig = new Config($config);
         $endpoint = $ossConfig->getOssEndpoint();
         $this->assertSame($config['internal'], $endpoint);
     }