+++ tests/TestCase.php
@@ -2,7 +2,6 @@
 
 namespace AlphaSnow\AliyunOss\Tests;
 
-use AlphaSnow\AliyunOss\AliyunOssClient;
 use AlphaSnow\AliyunOss\ServiceProvider;
 use Orchestra\Testbench\TestCase as BaseTestCase;
 
@@ -21,6 +20,6 @@ class TestCase extends BaseTestCase
 
     protected function getPackageAliases($app)
     {
-        return ['AliyunOssClient' => AliyunOssClient::class];
+        return [];
     }
 }