+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## [2.9.0](https://github.com/alphasnow/aliyun-oss-laravel/tree/2.8.0) (2021-08-16)
+[Show detail](https://github.com/alphasnow/aliyun-oss-laravel/compare/2.8.0...2.9.0)
+
+* Remove the facade of OssClient
+* Rename the name of base class
+* Fix phpunit test
+
 ## [2.8.0](https://github.com/alphasnow/aliyun-oss-laravel/tree/2.8.0) (2021-06-10)
 [Show detail](https://github.com/alphasnow/aliyun-oss-laravel/compare/2.7.0...2.8.0)